GPS working. Write & read checked. Button functionalities not implemented completely

+++ app/src/main/java/com/example/midel/stepper/SlotWalk.java
@@ -5,13 +5,13 @@ import com.google.android.gms.maps.model.LatLng;
 import java.io.Serializable;
 
 public class SlotWalk implements Serializable {
-    private float mAltitude;
+    private double mAltitude;
     private float mDistance;
     private LatLng mLocation;
     private long mSteps;
     private float mTime;
 
-    public SlotWalk(float aAltitude, float aDistance, LatLng aLocation, long aSteps, float aTime){
+    public SlotWalk(double aAltitude, float aDistance, LatLng aLocation, long aSteps, float aTime){
         mAltitude=aAltitude;
         mDistance=aDistance;
         mLocation=aLocation;
@@ -19,7 +19,7 @@ public class SlotWalk implements Serializable {
         mTime=aTime;
     }
 
-    public float getAltitude(){ return mAltitude;}
+    public double getAltitude(){ return mAltitude;}
     public float getDistance(){ return mDistance;}
     public LatLng getLocation(){ return mLocation;}
     public long getSteps(){ return mSteps;}
